+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add set_min_fluid_temperature and set_max_fluid_temperature to the Borefield class (issue #271).
 - Add flow per borefield to flow classes (issue #408).
 - Average baseload fluid temperature (issue #418).
+- Explicit formulations of the multipole method (issue #425).
+- Convective resistance calculation with array inputs (issue #425).
 
 ## Changed